+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/data_pipeline_supervisor.ex
@@ -1,7 +1,7 @@
 #
 # This file is part of Astarte.
 #
-# Copyright 2020 Ispirata Srl
+# Copyright 2020 - 2025 SECO Mind Srl
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,10 +21,7 @@ defmodule Astarte.DataUpdaterPlant.DataPipelineSupervisor do
 
   alias Astarte.DataUpdaterPlant.ConsumersSupervisor
   alias Astarte.DataUpdaterPlant.AMQPEventsProducer
-  alias Astarte.DataUpdaterPlant.RPC.Handler
   alias Astarte.DataUpdaterPlant.Config
-
-  alias Astarte.RPC.Protocol.DataUpdaterPlant, as: Protocol
 
   def start_link(init_arg) do
     Supervisor.start_link(__MODULE__, init_arg, name: __MODULE__)
@@ -33,15 +30,33 @@ defmodule Astarte.DataUpdaterPlant.DataPipelineSupervisor do
   @impl true
   def init(_init_arg) do
     children = [
-      {Registry, [keys: :unique, name: Registry.MessageTracker]},
-      {Registry, [keys: :unique, name: Registry.DataUpdater]},
+      {Horde.Registry, [keys: :unique, name: Registry.MessageTracker, members: :auto]},
+      {Horde.Registry, [keys: :unique, name: Registry.DataUpdater, members: :auto]},
+      {Horde.Registry, [keys: :unique, name: Registry.DataUpdaterRPC, members: :auto]},
+      {Horde.Registry, [keys: :unique, name: Registry.AMQPDataConsumer, members: :auto]},
+      {Horde.Registry, [keys: :unique, name: Registry.VMQPluginRPC, members: :auto]},
+      {Horde.DynamicSupervisor,
+       [
+         name: Supervisor.MessageTracker,
+         strategy: :one_for_one,
+         restart: :transient,
+         members: :auto,
+         distribution_strategy: Horde.UniformDistribution
+       ]},
+      {Horde.DynamicSupervisor,
+       [
+         name: Supervisor.DataUpdater,
+         strategy: :one_for_one,
+         restart: :transient,
+         members: :auto,
+         distribution_strategy: Horde.UniformDistribution
+       ]},
       {ExRabbitPool.PoolSupervisor,
        rabbitmq_config: Config.amqp_producer_options!(),
        connection_pools: [Config.events_producer_pool_config!()]},
       AMQPEventsProducer,
       ConsumersSupervisor,
-      {Astarte.RPC.AMQP.Server, [amqp_queue: Protocol.amqp_queue(), handler: Handler]},
-      Astarte.RPC.AMQP.Client
+      Astarte.DataUpdaterPlant.RPC.Supervisor
     ]
 
     Supervisor.init(children, strategy: :rest_for_one)