+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - Unreleased
+### Changed
+- Rework RPC between AppEngine and DUP using Erlang's native clustering and
+  message-passing instead of AMQP queues:
+  [#1186](https://github.com/astarte-platform/astarte/pull/1186).
+  Fix [#699](https://github.com/astarte-platform/astarte/issues/699).
+- Rework RPC between AppEngine and VerneMQ using Erlang's native clustering and
+  message-passing instead of AMQP queues:
+  [#1194](https://github.com/astarte-platform/astarte/pull/1194).
+- Rework RPC between DUP and VerneMQ using Erlang's native clustering and
+  message-passing instead of AMQP queues:
+  [#1197](https://github.com/astarte-platform/astarte/pull/1197).
+
 ## [1.2.1-alpha.0] - 2025-04-10
 ### Changed
 - Update the docker-compose configuration to allow both physical and virtual devices